# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CineLive API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me:
    get:
      operationId: getMyProfile
      tags: [users]
      summary: Get the authenticated user's profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"

    patch:
      operationId: updateMyProfile
      tags: [users]
      summary: Update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileBody"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
...
  /users/history/{showId}:
    post:
      operationId: recordWatchHistory
      tags: [users]
      summary: Record or update watch history for a show
      parameters:
        - name: showId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecordHistoryBody"
      responses:
        "200":
          description: History recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WatchHistory"
...
components:
  schemas:
...
    Show:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        thumbnailUrl:
          type: ["string", "null"]
        videoUrl:
          type: ["string", "null"]
        price:
          type: number
        category:
          type: string
          enum: [now_showing, upcoming]
        genre:
          type: ["string", "null"]
        duration:
          type: ["integer", "null"]
        startTime:
          type: ["string", "null"]
        isLive:
          type: boolean
        isFeatured:
          type: boolean
        ticketCount:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - title
        - description
        - price
        - category
        - isLive
        - isFeatured
        - ticketCount
        - createdAt
...
    UserProfile:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        displayName:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        isAdmin:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - email
        - isAdmin
        - createdAt
    UpdateProfileBody:
      type: object
      properties:
        displayName:
          type: string
        avatarUrl:
          type: string
    WatchHistory:
      type: object
      properties:
        id:
          type: integer
        showId:
          type: integer
        show:
          $ref: "#/components/schemas/Show"
        progressSeconds:
          type: integer
        watchedAt:
          type: string
      required:
        - id
        - showId
        - show
        - progressSeconds
        - watchedAt
    RecordHistoryBody:
      type: object
      properties:
        progressSeconds:
          type: integer
      required:
        - progressSeconds
...